now go through door without pressing space; just both players need to be colliding. Repositioned pressure plate so it doesn't bug out with uneven surface when sinking

+++ Assets/Scripts/PressurePlateRotation/PressurePlate.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 public class PressurePlate : MonoBehaviour
 {
     public PressurePlateController parent;
-    public float distToSink = (float)0.2;
+    public float distToSink = 0.2f;
     public int animationTimeInFrames = 10;
 
     private List<GameObject> localColliders;
